WIP : Code re-factoring : Moving mtp operations.
