+++ inc/mtp.h
@@ -143,6 +143,10 @@ int mtp_push_event(mtp_ctx * ctx, uint32_t event, int nbparams, uint32_t * param
 
 void mtp_deinit_responder(mtp_ctx * ctx);
 
+int build_response(mtp_ctx * ctx, uint32_t tx_id, uint16_t type, uint16_t status, void * buffer,void * datain,int size);
+int check_and_send_USB_ZLP(mtp_ctx * ctx , int size);
+int parse_incomming_dataset(mtp_ctx * ctx,void * datain,int size,uint32_t * newhandle, uint32_t parent_handle, uint32_t storage_id);
+
 #define APP_VERSION "v1.1.0"
 
 #endif
